// src/generated/models/Dga_ict_budget_line_itemsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Dga_ict_budget_line_itemsdga_added_in_allocation = {
  1: 'No',
  2: 'Yes'
} as const;
export type Dga_ict_budget_line_itemsdga_added_in_allocation = keyof typeof Dga_ict_budget_line_itemsdga_added_in_allocation;
export const Dga_ict_budget_line_itemsdga_expense_type = {
  1: 'Capex',
  2: 'Opex'
} as const;
export type Dga_ict_budget_line_itemsdga_expense_type = keyof typeof Dga_ict_budget_line_itemsdga_expense_type;
export const Dga_ict_budget_line_itemsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Dga_ict_budget_line_itemsstatecode = keyof typeof Dga_ict_budget_line_itemsstatecode;
export const Dga_ict_budget_line_itemsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Dga_ict_budget_line_itemsstatuscode = keyof typeof Dga_ict_budget_line_itemsstatuscode;

export interface Dga_ict_budget_line_itemsBase {
  dga_account_group?: string;
  dga_added_in_allocation?: Dga_ict_budget_line_itemsdga_added_in_allocation;
  dga_budget_allocated?: number;
  dga_budget_recommended?: number;
  dga_budget_requested?: number;
  "dga_classification@odata.bind"?: string;
  dga_description?: string;
  dga_ebs_account_code?: string;
  dga_expense_type?: Dga_ict_budget_line_itemsdga_expense_type;
  dga_fusion_account_code?: string;
  "dga_ict_budget@odata.bind"?: string;
  dga_ict_budget_line_itemid: string;
  dga_name?: string;
  dga_total_budget_utilized?: number;
  dga_utilization_quarter_1?: number;
  dga_utilization_quarter_2?: number;
  dga_utilization_quarter_3?: number;
  dga_utilization_quarter_4?: number;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  "OwningBusinessUnit@odata.bind"?: string;
  statecode: Dga_ict_budget_line_itemsstatecode;
  statuscode?: Dga_ict_budget_line_itemsstatuscode;
  timezoneruleversionnumber?: number;
  "TransactionCurrencyId@odata.bind"?: string;
  utcconversiontimezonecode?: number;
}

export interface Dga_ict_budget_line_items extends Dga_ict_budget_line_itemsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  dga_added_in_allocationname?: string;
  dga_budget_allocated_base?: number;
  dga_budget_recommended_base?: number;
  dga_budget_requested_base?: number;
  dga_classificationname?: string;
  dga_expense_typename?: string;
  dga_ict_budgetname?: string;
  dga_total_budget_utilized_base?: number;
  dga_utilization_quarter_1_base?: number;
  dga_utilization_quarter_2_base?: number;
  dga_utilization_quarter_3_base?: number;
  dga_utilization_quarter_4_base?: number;
  exchangerate?: number;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  transactioncurrencyidname?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  dga_classification?: object;
  _dga_classification_value?: string;
  dga_ict_budget?: object;
  _dga_ict_budget_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  transactioncurrencyid?: object;
  _transactioncurrencyid_value?: string;
}